+++ src/frontend/js/utils/indirection/window.ts
@@ -1,2 +1,2 @@
 // Layer of indirection to facilitate testing of code that would use location facilities
-export const { history, location, matchMedia, scroll } = window;
+export const { confirm, history, location, matchMedia, scroll } = window;